Dynamically import the consentless scripts
